chore: vendored freeink-sdk を上流の現行版に更新する
取り込んでいたのは CrossDiTo 1.5.1 公開時（2026-08-21）のスナップショットで、
その後の3週間ぶんの修正が入っていなかった。実機で出ていた2つの症状は、どちらも
上流ですでに直っていた。

  3d90c8a 2026-08-20 Fix UC8279X4 display orientation to match hardware
  ffeaaa2 2026-08-20 Fix UC8279 grayscale plane inversion bug
  df4a1b7 2026-08-25 Fix ghosting by forcing power-off after AA refresh
  4a69a29 2026-09-09 feat: accept absolute planes on UltraChip X4s
  39606d5 2026-09-09 Fix e-ink display buffer tracking and optimize SPI writes

Uc8279X4Driver は 361 行から 795 行になっている。手元で実機から割り出した
「SHL=1 ＋ 行は順方向」は、上流の現行版と一致していた（stock convention,
hardware-confirmed upright）。

freeink-sdk/ は 1.5.1 の取り込みコミット以来こちらでは変更していないので、
まるごと差し替えた。取り込みの構成（LICENSE / NOTICE / libs のみ）は維持。
lucide サブモジュールは上流でも手元でも未取得で対称。

上流 HEAD: 9f4d3f9ca675e64cc9d616081f39a33cdefce57e (2026-09-10)

**実機未確認。** 3週間ぶんの変更が入るので、他の箇所への影響も確かめる必要がある。
確認用にシリアルログ付きビルドを CI の成果物に加えている（確認後に外す）。

+++ freeink-sdk/libs/ui/FreeInkUI/include/components/controls/progress-bar.h
@@ -26,7 +26,7 @@ void progressBar(Frame<MaxInteractions>& frame, Rect rect, const ProgressBarProp
   }
   if (props.max <= 0 || props.value <= 0) return;
   const int32_t clamped = props.value > props.max ? props.max : props.value;
-  int16_t fillWidth = static_cast<int16_t>((static_cast<int32_t>(rect.width) * clamped) / props.max);
+  int16_t fillWidth = static_cast<int16_t>((static_cast<int64_t>(rect.width) * clamped) / props.max);
   if (fillWidth < props.minFill) fillWidth = props.minFill > rect.width ? rect.width : props.minFill;
   if (fillWidth > 0) {
     frame.target().fill(Rect{rect.x, rect.y, fillWidth, rect.height}, props.fill);
